requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.0
chromadb>=0.4.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.8.0
httpx>=0.25.0
python-multipart>=0.0.6
jinja2>=3.1.0
aiofiles>=23.0.0